tfrag: Separate out debug unpacker from new unpacker
